+++ src/asset.rs
@@ -17,23 +17,23 @@ pub struct Asset {
     path: Text,
 
     /// The asset's _current_ media type.
-    ///
-    /// When the asset is exported as a file, its extension
-    /// will match this media type's preferred extension.
-    pub media_type: MediaType,
+    media_type: MediaType,
 
     /// The asset's raw contents
-    pub contents: AssetContents,
+    contents: Option<AssetContents>,
 }
 
 impl Asset {
     /// Returns a new asset with `path` and `contents`.
     pub fn new(path: Text, contents: Vec<u8>) -> Self {
-        // Try to convert the vector to UTF-8 bytes.
-        let contents = if let Ok(text) = str::from_utf8(&contents) {
-            AssetContents::Text(text.into())
+        let contents = if contents.is_empty() {
+            None
         } else {
-            AssetContents::Binary(contents)
+            // Try to convert the vector to UTF-8 bytes.
+            Some(match String::from_utf8(contents) {
+                Ok(text) => AssetContents::Textual(text.into()),
+                Err(e) => AssetContents::Binary(e.into_bytes()),
+            })
         };
 
         // Extract the media type from the path.
@@ -50,39 +50,71 @@ impl Asset {
     pub fn path(&self) -> &Text {
         &self.path
     }
+
+    /// Returns the asset's media type.
+    pub fn media_type(&self) -> &MediaType {
+        &self.media_type
+    }
+
+    /// Sets the asset's media type.
+    pub fn set_media_type(&mut self, media_type: MediaType) {
+        self.media_type = media_type;
+    }
+
+    /// Replaces the assets contents with `bytes` and `media_type`.
+    pub fn replace_with_bytes(&mut self, bytes: Vec<u8>, media_type: MediaType) {
+        self.contents = Some(AssetContents::Binary(bytes));
+        self.media_type = media_type;
+    }
+
+    /// Replaces the assets contents with `text` and `media_type`.
+    pub fn replace_with_text(&mut self, text: Text, media_type: MediaType) {
+        self.contents = Some(AssetContents::Textual(text));
+        self.media_type = media_type;
+    }
+
+    /// Returns the asset's contents as immutable bytes.
+    pub fn as_bytes(&self) -> &[u8] {
+        match self.contents.as_ref() {
+            Some(AssetContents::Binary(bytes)) => bytes,
+            Some(AssetContents::Textual(text)) => text.as_bytes(),
+            None => &[],
+        }
+    }
+
+    /// Returns the assets contents as immutable text.
+    ///
+    /// If the asset is empty or contains non-textual data,
+    /// this function will fail.
+    pub fn as_text(&self) -> Result<&Text, AssetError> {
+        match self.contents.as_ref() {
+            Some(AssetContents::Textual(text)) => Ok(text),
+            _ => Err(AssetError::NonTextual),
+        }
+    }
+
+    /// Returns the asset's contents as mutable bytes.
+    ///
+    /// If the asset is empty, this function will fail.
+    ///
+    /// If the asset contains text, this function will fail:
+    /// All assets can be _represented_ [as bytes](Self::as_bytes),
+    /// but it would be unsafe to modify a textual asset's bytes
+    /// in place, since the resulting bytes may no longer
+    /// represent valid text.
+    pub fn as_mut_bytes(&mut self) -> Result<&mut Vec<u8>, AssetError> {
+        match &mut self.contents {
+            Some(AssetContents::Binary(bytes)) => Ok(bytes),
+            _ => Err(AssetError::NonBinary),
+        }
+    }
 }
 
 /// Raw contents of an [Asset].
 #[derive(Clone, Debug)]
-pub enum AssetContents {
+enum AssetContents {
     Binary(Vec<u8>),
-    Text(Text),
-}
-
-impl AssetContents {
-    /// Returns the contents as raw bytes.
-    pub fn as_bytes(&self) -> &[u8] {
-        match &self {
-            AssetContents::Binary(data) => data,
-            AssetContents::Text(text) => text.as_bytes(),
-        }
-    }
-
-    /// Returns the contents as mutable bytes.
-    pub fn try_as_mut_bytes(&mut self) -> Result<&mut Vec<u8>, AssetError> {
-        match self {
-            AssetContents::Binary(bytes) => Ok(bytes),
-            _ => Err(AssetError::NonBinary),
-        }
-    }
-
-    /// Returns the contents as mutable text.
-    pub fn try_as_mut_text(&mut self) -> Result<&mut Text, AssetError> {
-        match self {
-            AssetContents::Text(text) => Ok(text),
-            _ => Err(AssetError::NonTextual),
-        }
-    }
+    Textual(Text),
 }
 
 /// A thing that processes [Asset]s.
@@ -91,7 +123,7 @@ pub trait ProcessesAssets {
     fn process(&self, asset: &mut Asset) -> Result<(), AssetError>;
 }
 
-#[derive(Debug)]
+#[derive(Debug, PartialEq, Eq)]
 pub enum AssetError {
     /// An asset contained data that wasn't text.
     NonTextual,
@@ -113,29 +145,21 @@ mod tests {
 
     #[test]
     fn creates_assets() {
-        let mut markdown_asset = Asset::new("story.md".into(), "Hello, world!".as_bytes().to_vec());
+        let markdown_asset = Asset::new("story.md".into(), "Hello, world!".as_bytes().to_vec());
         assert_eq!("story.md", markdown_asset.path());
-        assert_eq!(MediaType::Markdown, markdown_asset.media_type);
-        assert_eq!(
-            "Hello, world!".as_bytes(),
-            markdown_asset.contents.as_bytes()
-        );
-        assert_eq!(
-            "Hello, world!",
-            markdown_asset.contents.try_as_mut_text().unwrap()
-        );
+        assert_eq!(&MediaType::Markdown, markdown_asset.media_type());
+        assert_eq!(b"Hello, world!", markdown_asset.as_bytes());
+        assert_eq!("Hello, world!", markdown_asset.as_text().unwrap());
 
         let binary_asset = Asset::new("data.dat".into(), (-1337i16).to_le_bytes().to_vec());
         assert_eq!("data.dat", binary_asset.path());
         assert_eq!(
-            MediaType::Unknown {
+            &MediaType::Unknown {
                 extension: ["dat".into()]
             },
-            binary_asset.media_type
+            binary_asset.media_type()
         );
-        assert_eq!(
-            &(-1337i16).to_le_bytes().to_vec(),
-            binary_asset.contents.as_bytes()
-        );
+        assert_eq!(&(-1337i16).to_le_bytes().to_vec(), binary_asset.as_bytes(),);
+        assert_eq!(Err(AssetError::NonTextual), binary_asset.as_text());
     }
 }