+++ src/asset/markdown.rs
@@ -13,7 +13,16 @@ pub struct MarkdownProcessor {}
 
 impl ProcessesAssets for MarkdownProcessor {
     fn process(&self, asset: &mut super::Asset) -> Result<(), AssetError> {
-        let text = asset.contents.try_as_mut_text()?;
+        if *asset.media_type() != MediaType::Markdown {
+            tracing::debug!(
+                "skipping asset {}: not markdown {}",
+                asset.path(),
+                asset.media_type().name()
+            );
+            return Ok(());
+        }
+
+        let text = asset.as_text()?;
 
         // Compile markdown into an abstract syntax tree.
         let ast = markdown::to_mdast(text, &markdown::ParseOptions::default())?;
@@ -23,8 +32,7 @@ impl ProcessesAssets for MarkdownProcessor {
         compile_ast_node(None, &ast, &mut compiled_html);
 
         // Update the asset's contents and target extension.
-        *text.to_mut() = compiled_html;
-        asset.media_type = MediaType::Html;
+        asset.replace_with_text(compiled_html.into(), MediaType::Html);
         Ok(())
     }
 }
@@ -266,7 +274,9 @@ mod tests {
 
         assert_eq!(
             "<h1 id=\"header-1\">Header 1</h1><p>Body</p><Blockquote><p>Quotation in <strong>bold</strong> and <em>italics</em>.</p></Blockquote>",
-            markdown_asset.contents.try_as_mut_text().unwrap()
+            markdown_asset.as_text().unwrap()
         );
+
+        assert_eq!(&MediaType::Html, markdown_asset.media_type());
     }
 }