+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This file contains a chronological list of all major changes made to this crate.
 - Add markdown asset processor
 - Add grass crate to project and implement SCSS to css conversion (#5)
 - Implement Lanczos-based image resizing processor
+- Encapsulate details of assets' contents inside the asset module.
 
 ### 🐛 Fixes